Rerun all tests and fix 3 and 9 to output valid format and fix tester.sh to include correct answer for problem 16

+++ John/9.py
@@ -9,4 +9,4 @@ for i in range(1,1001):
     for j in range(1,1001):
         hyp = math.sqrt(i**2+j**2)
         if hyp.is_integer() and i+j+hyp == 1000 and j < i:
-            print(i*j*hyp)
+            print(int(i*j*hyp))
